# docs/openapi/gsf-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GSF API",
    "version": "0.1.1",
    "description": "Public HTTP surface of GSF, served by the Next.js route handlers. Generated by `pnpm openapi` — do not edit by hand."
  },
  "paths": {
...
    "/api/terms": {
      "get": {
        "description": "termsApi.list — one page of Terms and their per-card count breakdowns,\nzone-scoped for viewers. `skip`/`limit` pass straight through.",
        "operationId": "get_api_terms",
        "parameters": [
          {
            "description": "Case-insensitive substring filter on the term name.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring filter on the term name.",
              "title": "Q"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermsPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Terms",
        "tags": ["terms"],
        "x-gsf-source": "frontend/app/api/terms/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/terms/{term_id}": {
      "get": {
        "description": "termsApi.getById — fetch one Term node, zone-scoped for viewers.\n\nThis returns a single object (`{ data: TermDetail }`), not a list, so it\ncan't reuse proxyToBackend's empty-zone-viewer shortcut (which returns the\nlist-shaped `{ data: [], count: 0 }`) — a zero-zone viewer gets a 404\ninstead, matching what the backend returns for an out-of-zone/missing term.",
        "operationId": "get_api_terms_term_id_",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TermDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "termsApi.getById — fetch one Term node, zone-scoped for viewers.",
        "tags": ["terms"],
        "x-gsf-source": "frontend/app/api/terms/[term_id]/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      },
      "patch": {
        "description": "Updating a Term name invalidates cached SqlAttribute description suggestions in backend.",
        "operationId": "patch_api_terms_term_id_",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Term_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Term",
        "tags": ["terms"],
        "x-gsf-source": "frontend/app/api/terms/[term_id]/route.ts",
        "x-gsf-permissions": {
          "catalog": ["edit"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/terms/{term_id}/column-attributes": {
      "get": {
        "description": "termsApi.getColumnAttributes — ColumnAttribute nodes for one term (with\nprimary/referenced columns), zone-scoped for viewers.",
        "operationId": "get_api_terms_term_id_column_attributes",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListResponse_ColumnAttribute_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Term Column Attributes By Id",
        "tags": ["terms"],
        "x-gsf-source": "frontend/app/api/terms/[term_id]/column-attributes/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/terms/{term_id}/column-attributes/{attr_id}": {
      "patch": {
        "description": "Updating a ColumnAttribute changes only name/description; backend refreshes\nthe corresponding semantic VDB embedding.",
        "operationId": "patch_api_terms_term_id_column_attributes_attr_id_",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          },
          {
            "description": "ColumnAttribute id, not a SqlAttribute id.",
            "in": "path",
            "name": "attr_id",
            "required": true,
            "schema": {
              "description": "ColumnAttribute id, not a SqlAttribute id.",
              "title": "Attr Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnAttributeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ColumnAttribute_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Column Attribute",
        "tags": ["terms"],
        "x-gsf-source": "frontend/app/api/terms/[term_id]/column-attributes/[attr_id]/route.ts",
        "x-gsf-permissions": {
          "catalog": ["edit"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/terms/{term_id}/sql-attributes": {
      "get": {
        "description": "termsApi.getSqlAttributes — SqlAttribute nodes for one term, zone-scoped for viewers.",
        "operationId": "get_api_terms_term_id_sql_attributes",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListResponse_SqlAttribute_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Term Sql Attributes By Id",
        "tags": ["terms"],
        "x-gsf-source": "frontend/app/api/terms/[term_id]/sql-attributes/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColumnAttributeUpdate": {
        "properties": {
          "certified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certified"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sample_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Values"
          }
        },
        "title": "ColumnAttributeUpdate",
        "type": "object"
      },
...
      "DataResponse_ColumnAttribute_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ColumnAttribute"
          }
        },
        "required": ["data"],
        "title": "DataResponse[ColumnAttribute]",
        "type": "object"
      },
...
      "DataResponse_Term_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Term"
          }
        },
        "required": ["data"],
        "title": "DataResponse[Term]",
        "type": "object"
      },
      "DataResponse_TermDetail_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TermDetail"
          }
        },
        "required": ["data"],
        "title": "DataResponse[TermDetail]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PagedListResponse_ColumnAttribute_": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ColumnAttribute"
            },
            "title": "Data",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["data", "count", "total"],
        "title": "PagedListResponse[ColumnAttribute]",
        "type": "object"
      },
      "PagedListResponse_SqlAttribute_": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SqlAttribute"
            },
            "title": "Data",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["data", "count", "total"],
        "title": "PagedListResponse[SqlAttribute]",
        "type": "object"
      },
...
      "TermCountEntry": {
        "additionalProperties": true,
        "description": "``{term_id, count}`` — one per-term breakdown row. Zeroes are omitted.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "term_id": {
            "title": "Term Id",
            "type": "string"
          }
        },
        "required": ["term_id", "count"],
        "title": "TermCountEntry",
        "type": "object"
      },
...
      "TermListItem": {
        "additionalProperties": true,
        "description": "A Term as it appears in the paged ``/terms`` list.",
        "properties": {
          "certification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certification"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "description_certified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Certified"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "name_certified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Certified"
          },
          "schema_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Names"
          },
          "synonyms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synonyms"
          },
          "zones": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ZoneChip"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zones"
          }
        },
        "required": ["id"],
        "title": "TermListItem",
        "type": "object"
      },
      "TermsPageResponse": {
        "additionalProperties": true,
        "description": "One page of the Terms list with its per-term count breakdowns.\n\n``total`` counts every matching term; the count lists cover only the\nterms on this page when a limit was given.",
        "properties": {
          "column_attribute_counts": {
            "items": {
              "$ref": "#/components/schemas/TermCountEntry"
            },
            "title": "Column Attribute Counts",
            "type": "array"
          },
          "related_counts": {
            "items": {
              "$ref": "#/components/schemas/TermCountEntry"
            },
            "title": "Related Counts",
            "type": "array"
          },
          "sql_attribute_counts": {
            "items": {
              "$ref": "#/components/schemas/TermCountEntry"
            },
            "title": "Sql Attribute Counts",
            "type": "array"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/TermListItem"
            },
            "title": "Terms",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "terms",
          "total",
          "column_attribute_counts",
          "sql_attribute_counts",
          "related_counts"
        ],
        "title": "TermsPageResponse",
        "type": "object"
      },
...
      "TermUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "description_certified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Certified"
          },
          "name": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "name_certified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Certified"
          }
        },
        "title": "TermUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  }
}